Take in a run_id in vm_run_telemetry and do not copy page_sets from GoogleStorage for every run.
(SkipBuildbotRuns)

Review URL: https://codereview.chromium.org/16831009

git-svn-id: http://skia.googlecode.com/svn/buildbot@9674 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ compute_engine_scripts/telemetry/telemetry_slave_scripts/vm_run_telemetry.sh
@@ -14,12 +14,12 @@
 if [ $# -ne 4 ]; then
   echo
   echo "Usage: `basename $0` 1 skpicture_printer" \
-    "--skp-outdir=/home/default/storage/skps/ rmistry"
+    "--skp-outdir=/home/default/storage/skps/ rmistry2013-05-24.07-34-05"
   echo
   echo "The first argument is the slave_num of this telemetry slave."
   echo "The second argument is the telemetry benchmark to run on this slave."
   echo "The third argument are the extra arguments that the benchmark needs."
-  echo "The fourth argument is the user who triggered the run."
+  echo "The fourth argument is the runid (typically requester + timestamp)."
   echo
   exit 1
 fi
@@ -27,13 +27,11 @@ fi
 SLAVE_NUM=$1
 TELEMETRY_BENCHMARK=$2
 EXTRA_ARGS=$3
-REQUESTOR=$4
+RUN_ID=$4
 
 source vm_utils.sh
 
-TIMESTAMP=`date "+%m-%d-%Y.%T"`
-WORKER_FILE=$TELEMETRY_BENCHMARK.$REQUESTOR.$TIMESTAMP
-create_worker_file $WORKER_FILE
+create_worker_file TELEMETRY_${RUN_ID}
 
 source vm_setup_slave.sh
 
@@ -42,9 +40,11 @@ source vm_setup_slave.sh
 # gsutil cp gs://chromium-skia-gm/telemetry/webpages_archive/slave$SLAVE_NUM/* \
 #   /home/default/storage/webpages_archive/
 
-# Clean and create the skp output directory.
-rm -rf /home/default/storage/skps
-mkdir -p /home/default/storage/skps/
+if [ "$TELEMETRY_BENCHMARK" == "skpicture_printer" ]; then
+  # Clean and create the skp output directory.
+  rm -rf /home/default/storage/skps
+  mkdir -p /home/default/storage/skps/
+fi
 
 for page_set in /home/default/storage/page_sets/*; do
   if [[ -f $page_set ]]; then
@@ -79,4 +79,4 @@ if [ "$TELEMETRY_BENCHMARK" == "skpicture_printer" ]; then
     gs://chromium-skia-gm/telemetry/skps/slave$SLAVE_NUM/
 fi
 
-delete_worker_file $WORKER_FILE
+delete_worker_file TELEMETRY_${RUN_ID}
